Dev new branch h

+++ library_management.client/src/app/admin/add-book-common/add-book-common.component.ts
@@ -16,6 +16,8 @@ declare var
   styleUrls: ['./add-book-common.component.css']
 })
 export class AddBookCommonComponent {
+  book$ = this.exploreService.book$;
+
   /*@ViewChild('exampleModalCenter') modal: any;*/
   @Input() stepperIndex: number = 0;
 
@@ -37,10 +39,11 @@ export class AddBookCommonComponent {
   captureImg:any=""
   getPicture(event: any) {
     this.bookForm.patchValue({
-      img: ''
+      img: event
     });
 
-    this.captureImg = event;
+    //this.captureImg = event;
+    this.exploreService.setImg(event);
     console.log("hello",this.capturedImage);
   }
 
@@ -100,6 +103,37 @@ export class AddBookCommonComponent {
       }
     );
 
+
+    //this.book$.subscribe(book => {
+    //  console.log(book);
+    //});
+
+
+    this.exploreService.book$.subscribe(book => {
+      if (book) {
+        this.bookForm.patchValue({
+          bookName: book.bookName,
+          authorName: book.authorName,
+          img: book.img,
+          description: book.description
+        });
+        this.selectedBook = book.bookName;
+      }
+    });
+
+    var api = "https://www.googleapis.com/books/v1/volumes?q=default";
+
+    //setTimeout(() => {
+    fetch(api)
+      .then(response => response.json())
+      .then(data => {
+        console.log(data);
+        this.listOfBooks = data.items;
+
+      })
+      .catch(error => {
+        console.error('Error fetching books:', error);
+      });
   }
 
   openWebcam() {
@@ -211,48 +245,25 @@ export class AddBookCommonComponent {
   }
 
   decrement(index: number) {
-    if (this.counterValue > 1) {
-      this.counterValue--;
-      this.bookForm.patchValue({
-        qty: this.counterValue,
-      });
-      const qrArray = this.bookForm.get('qr') as FormArray;
-      qrArray.removeAt(this.counterValue);
-      console.log(this.bookForm);
+    if (this.qrCodes.length > 1) {
+      if (this.counterValue > 1) {
+        this.counterValue--;
+        this.bookForm.patchValue({
+          qty: this.counterValue,
+        });
+        const qrArray = this.bookForm.get('qr') as FormArray;
+        qrArray.removeAt(this.counterValue);
+        console.log(this.bookForm);
+      }
+
+
+
+      this.exploreService.removeQrCode(index);
     }
-
-
-
-    this.exploreService.removeQrCode(index);
     console.log("array size", this.qrCodes);
   }
 
-  //updateQuantityValue(event: any) {
-  //  const newValue = parseInt(event.target.value, 10);
-  //  const deviceIdArray = this.addDeviceForm.get('deviceId') as FormArray;
-
-  //  if (!isNaN(newValue)) {
-  //    const currentValue = deviceIdArray.length;
-
-  //    if (newValue > currentValue) {
-  //      const elementsToAdd = newValue - currentValue;
-  //      for (let i = 1; i <= elementsToAdd; i++) {
-  //        this.pushValueIntoDeviceId('CGI-MOU-' + (this.laststoredcgi + i));
-  //      }
-  //    }
-
-  //    else if (newValue < currentValue) {
-  //      const elementsToRemove = currentValue - newValue;
-  //      for (let i = 0; i < elementsToRemove; i++) {
-  //        deviceIdArray.removeAt(deviceIdArray.length - 1);
-  //      }
-  //    }
-
-  //    this.counterValue = newValue;
-  //    this.addDeviceForm.get('qty')?.setValue(newValue);
-  //  }
-  //}
-
+ 
   stepperIncrement() {
     this.stepperIndex++;
     this.exploreService.setaddBookPage(this.stepperIndex);
@@ -260,7 +271,7 @@ export class AddBookCommonComponent {
 
   stepperDecrement() {
     this.stepperIndex--;
-    this.exploreService.setaddBookPage(this.stepperIndex);
+    this.exploreService.setaddBookPage(this.stepperIndex); 
   }
 
   getBooks(event: any, type: string) {
@@ -322,6 +333,16 @@ export class AddBookCommonComponent {
 
   testing(event: any) {
     console.log("hello", event.target);
+    
+
+    this.exploreService.setBook({
+      bookName: this.selectedBook.volumeInfo.title,
+      authorName: this.selectedBook.volumeInfo.authors,
+      img: this.selectedBook.volumeInfo.imageLinks.smallThumbnail,
+      description: this.selectedBook.volumeInfo.description,
+      ISBN: this.selectedBook.volumeInfo.industryIdentifiers.find((id:any) => id.type === "ISBN_13").identifier,
+    });
+
     if (this.selectedBook) {
       this.bookForm.patchValue({
         bookName: this.selectedBook.volumeInfo.title,
@@ -330,9 +351,12 @@ export class AddBookCommonComponent {
         description: this.selectedBook.volumeInfo.description,
       });
       this.captureImg = '';
+      console.log("book selected", this.bookForm);
+      this.selectedBook = this.selectedBook.volumeInfo.title;
     }
+
+
     
-    console.log(this.bookForm);
     return;
   }
 
@@ -378,5 +402,16 @@ export class AddBookCommonComponent {
 
     this.router.navigate(['/admin/issue-mobile-scanner'], navigationExtras);
   }
+  routeBasedOnScreenSize() {
+      if (window.innerWidth <= 765) {
+        this.router.navigate(['admin/add-book-scanner']);
+      } else {
+        this.openWebcam();
+
+      }
+  }
+ 
 
 }
+
+
